requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
python-dotenv>=1.0.0
requests>=2.31.0
folium>=0.17.0
pandas>=2.2.0
shapely>=2.0.0
